Antenna.js full translated from Matlab
The preceding file in the repo had some non-working code, not still translated from Matlab. The version I propose is a full translation of functions antennafunccircular, array_factorcir and trapezoidalcir, found here:
http://www3.ntu.edu.sg/home/epnsugan/index_files/CEC11-RWP/CEC11-RWP.htm

I have tested it with several inputs and the results are the same than the produced by the original functions. However, a more extensive test is mandatory. It also remains to add the license and the corresponding credits and a deep format revision.